Youtube-Summarizer-master/requirements.txt
bert_score==0.3.13
nltk==3.8.1
numpy==1.26.3
openai_whisper==20231117
pytextrank==3.2.5
pytube==15.0.0
pytube3==9.6.4
# rouge==1.0.1
py_rouge==1.1
spacy==3.7.2
streamlit==1.28.2
torch==2.1.1
transformers==4.35.2
youtube_transcript_api==0.6.1
